backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
